refactor: Rename transformer to source-spec-transformer
This paves the way for having transformers for other sections of the
spec processing pipeline (e.g. tweaking code samples in the final
`merged_code_samples_specs` step).
