+++ package.json
@@ -22,7 +22,7 @@
     "test": "vitest run",
     "test:all": "pnpm test",
     "test:watch": "vitest",
-    "transform": "node src/index.js"
+    "transform:source_specs": "node src/index.js --source_specs"
   },
   "dependencies": {
     "js-yaml": "^4.1.0"
