+++ src/source-spec-transformer.js
